requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
dnspython==2.1.0
emoji==1.2.0
idna==2.10
nmap==0.0.1
phonenumbers==8.12.17
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.51.1
urllib3==1.26.3
